pyproject.toml
[project]
name = "homalos"
dynamic = ["version"]
description = "基于Python的期货量化交易系统"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Donny", email = "donnymoving@gmail.com" }]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Office/Business :: Financial :: Investment",
    "Natural Language :: Chinese (Simplified)",
    "Typing :: Typed",
]
dependencies = [
    "hatch>=1.14.1",
    "hatchling>=1.27.0",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "meson>=1.8.1",
    "meson-python>=0.18.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pip>=25.1.1",
    "polars>=1.30.0",
    "polib>=1.2.0",
    "pybind11>=2.13.6",
    "pybind11-stubgen>=2.5.4",
    "pyyaml>=6.0.2",
    "ta-lib>=0.6.3",
    "torch>=2.7.0",
    "tzlocal>=5.3.1",
]

keywords = ["quant", "quantitative", "trading", "homalos", "investment"]

[project.urls]
"Homepage" = "https://github.com/Homalos"
"Documentation" = "https://github.com/Homalos/Homalos/README.md"
"Repository" = "https://github.com/Homalos/Homalos.git"
"Issues" = "https://github.com/Homalos/Homalos/issues"
"Changes" = "https://github.com/Homalos/Homalos/CHANGELOG.md"
"Source" = "https://github.com/Homalos/Homalos"

[build-system]
requires = [
    "hatchling",
    "meson-python>=0.17.1",
    "meson>=1.7.0",
    "pybind11>=2.13.6",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/__version__.py"
pattern = "__version__ = ['\"](?P<version>[^'\"]+)['\"]"

[tool.hatch.build.targets.sdist]
include = ["src*"]

[tool.hatch.build.targets.wheel]
packages = ["src*"]

[tool.hatch.envs.default]
dependencies = [
  "pytest>=7.4.0",
  "ruff>=0.3.0",  # 静态代码分析工具
  "mypy>=1.8.0",    # 类型检查工具
]

[tool.mesonpy]
build_dir = "build"
args = ['--vsenv']

[tool.mesonpy.distribution]
include = [
    "**/*.dll",
    "**/*.so",
    "**/*.pyd",
    "**/*.h",
    "**/*.cpp",
    "**/*.lib",
    "**/*.a",
    "**/*.framework/**",  # 添加Mac框架文件夹支持
]

preserve-path = true

[tool.ruff]
target-version = "py312"
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true
